+++ next.config.js
@@ -2,6 +2,9 @@
 module.exports = {
     reactStrictMode: false,
     images: {
-        domains: ["image.tmdb.org","cdn.arabsstock.com"],
+        domains: ["image.tmdb.org", "cdn.arabsstock.com"],
+    },
+    experimental: {
+        appDir: true,
     },
 };
