+++ package.json
@@ -1,26 +1,24 @@
 {
-  "private": true,
-  "scripts": {
-    "dev": "next dev",
-    "build": "next build",
-    "start": "next start"
-  },
-  "dependencies": {
-    "@reduxjs/toolkit": "^1.9.2",
-    "axios": "^1.3.4",
-    "next": "latest",
-    "next-redux-wrapper": "^8.1.0",
-    "react": "18.2.0",
-    "react-dom": "18.2.0",
-    "react-redux": "^8.0.5"
-  },
-  "devDependencies": {
-    "@types/node": "18.14.1",
-    "@types/react": "18.0.28",
-    "@types/react-dom": "18.0.6",
-    "autoprefixer": "^10.4.12",
-    "postcss": "^8.4.18",
-    "tailwindcss": "^3.2.4",
-    "typescript": "4.9.5"
-  }
+    "private": true,
+    "scripts": {
+        "dev": "next dev",
+        "build": "next build",
+        "start": "next start"
+    },
+    "dependencies": {
+        "axios": "^1.3.4",
+        "next": "latest",
+        "react": "latest",
+        "react-dom": "latest",
+        "eslint-config-next": "latest"
+    },
+    "devDependencies": {
+        "@types/node": "18.14.1",
+        "@types/react": "18.0.28",
+        "@types/react-dom": "18.0.6",
+        "autoprefixer": "^10.4.12",
+        "postcss": "^8.4.18",
+        "tailwindcss": "^3.2.4",
+        "typescript": "4.9.5"
+    }
 }
