+++ pages/api/test.ts
@@ -1,0 +1,7 @@
+import { NextApiRequest, NextApiResponse } from "next";
+
+export default function handler(req: NextApiRequest, res: NextApiResponse) {
+    return res.status(200).json({
+        message: "Working Success!",
+    });
+}
