--- store.ts
@@ -1,18 +1,0 @@
-import { Action, configureStore, ThunkAction } from "@reduxjs/toolkit";
-import { createWrapper } from "next-redux-wrapper";
-import { authSlice } from "./slices/authSlice";
-
-export const store = configureStore({
-    reducer: {
-        [authSlice.name]: authSlice.reducer
-    },
-    devTools: true,
-});
-
-const makeStore = () => store;
-
-export type AppStore = ReturnType<typeof makeStore>;
-export type AppState = ReturnType<AppStore["getState"]>;
-export type AppThunk<ReturnType = void> = ThunkAction<ReturnType, AppState, unknown, Action>;
-
-export const wrapper = createWrapper<AppStore>(makeStore);
