+++ utils.ts
@@ -1,0 +1,1 @@
+export const genTitle = (prefix: string) => `${prefix} | mFlux - Next`;
